Added Potions and clickable consumables in Inventory
Health Potion is now added. When they are clicked in the inventory, they heal the player and are then removed from the inventory.

+++ 2D-RPG/Assets/Scripts/Other/DestroyObject.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class DestroyObject : MonoBehaviour
 {
-    public int timeToDestroy;
+    public float timeToDestroy;
 
 	void Start ()
     {
